requirements.txt
flask==3.1.3
flask-cors==5.0.1
gunicorn
onnxruntime
numpy
pillow
